refactor: move Domain, Range to types.ts and change the way to override default options
