+++ __tests__/unit/scales/base.spec.ts
@@ -1,7 +1,6 @@
-import { Base, DEFAULT_OPTIONS } from '../../../src/scales/base';
-import { BaseOptions } from '../../../src/types';
+import { Base } from '../../../src/scales/base';
+import { BaseOptions, Domain, Range } from '../../../src/types';
 import { ticks } from '../../../src/tick-method/basic';
-import { Domain, Range } from '../../../src/utils/type';
 
 class Scale extends Base<BaseOptions> {
   // eslint-disable-next-line class-methods-use-this
@@ -50,7 +49,7 @@ describe('Scale', () => {
   });
 
   test('Scale({}, defaults) set defaults', () => {
-    const s = new Scale({}, { ...DEFAULT_OPTIONS, domain: [0, 20] });
+    const s = new Scale({}, { domain: [0, 20] });
 
     // @ts-ignore
     expect(s.options.domain).toEqual([0, 20]);
