+++ src/types.ts
@@ -1,6 +1,12 @@
 export type TickMethod<T> = (options?: T) => any[];
 
-export type BaseOptions<R = any, D = any, T = any> = {
+/**
+ * 所有比例尺选项的默认类型
+ * D：定义域元素的类型
+ * R：值域元素的类型
+ * T：tickMethod 配置项的类型
+ */
+export type BaseOptions<D = any, R = any, T = any> = {
   /** 当需要映射的值不合法的时候，返回的值 */
   unknown?: any;
   /** 值域，默认为 [0, 1] */
@@ -16,6 +22,15 @@ export type BaseOptions<R = any, D = any, T = any> = {
   /** 计算 ticks 的算法 */
   tickMethod?: TickMethod<T>;
 };
+
+/** 获得比例尺选项中定义域元素的类型 */
+export type Domain<O extends BaseOptions> = O['domain'][number];
+
+/** 获得比例尺选项中值域元素的类型 */
+export type Range<O extends BaseOptions> = O['range'][number];
+
+/** 获得比例尺选项中 unknown 的类型 */
+export type Unknown<O extends BaseOptions> = O['unknown'];
 
 export type IdentityOptions = BaseOptions<number, number>;
