--- src/utils/type.ts
@@ -1,7 +1,0 @@
-import { BaseOptions } from '../types';
-
-export type Domain<O extends BaseOptions> = O['domain'][number];
-
-export type Range<O extends BaseOptions> = O['range'][number];
-
-export type Unknown<O extends BaseOptions> = O['unknown'];
